fix 2 extra blocks & add function into common.js #683
