+++ blocks/recommendations/recommendations.js
@@ -1,4 +1,4 @@
-import { createElement, getTextLabel } from '../../scripts/common.js';
+import { createElement, getArticleTags, getTextLabel } from '../../scripts/common.js';
 import {
   getAllArticles,
   getLimit,
@@ -16,7 +16,7 @@ const readNowText = getTextLabel('READ NOW');
 
 export default async function decorate(block) {
   const limit = Number(getLimit(block));
-  const category = getMetadata('category');
+  const category = await getArticleTags('categories') || getMetadata('category');
   const allArticles = await getAllArticles();
 
   const recommendedArticles = allArticles.filter((e) => e.category === category);
