+++ scripts/common.js
@@ -586,3 +586,47 @@ export const clearElementAttributes = (element) => {
 
   return element;
 };
+
+// magazine common functions
+/**
+ * Extracts the values from an array of objects and returns an array of values
+ * example: [{ key: 'value' }] => ['value']
+ * @param {array} array - An array of objects
+ * @returns {array} An array of values
+ */
+function getValuesFromObjectsArray(array = []) {
+  if (!Array.isArray(array) || array.length === 0) return [];
+  return array.map((item) => Object.values(item)[0]);
+}
+
+/**
+ * Extracts the matching tags from an array of tags and an array of article tags
+ * and returns a string of matching tags
+ * @param {array} tags - An array of tags from the JSON file
+ * @param {array} articleTags - An array of article:tags
+ * @returns {string} A string of matching tags
+ */
+function getMetadataFromTags(tags, articleTags) {
+  if (!tags || !articleTags) {
+    return '';
+  }
+
+  const matchingTags = [...articleTags]
+    .filter((tag) => tags.includes(tag.content))
+    .map((tag) => tag.content);
+  return matchingTags && matchingTags?.length > 0 ? matchingTags.join(', ') : '';
+}
+
+/**
+ * Get the article tags from the JSON file and the article tags from the document
+ * and return the matching tags
+ * @param {string} tagType - The type of tag to get such as 'categories', 'topics' or 'trucks'
+ * @returns {string} A string of matching tags
+ */
+export async function getArticleTags(tagType) {
+  const articleTags = document.head.querySelectorAll('meta[property="article:tag"]') || [];
+  const tagItems = await getJsonFromUrl('/magazine/articles/tags.json') || null;
+  const tags = tagItems && tagItems[tagType]
+    && getValuesFromObjectsArray(tagItems[tagType].data);
+  return getMetadataFromTags(tags, articleTags);
+}
